… NOTICE.txt
… go.sum