<!-- Thanks for sending a pull request!

If this is your first contribution, please review and sign our contributor agreement -
https://www.elastic.co/contributor-agreement.

Guidelines:
 - Prefer small PRs, and split changes into multiple logical commits where they must
   be delivered in a single PR.
 - If the PR is incomplete and not yet ready for review, open it as a Draft.
 - Once the PR is marked ready for review it is expected to pass all tests and linting,
   and you should not force-push any changes.

See also https://github.com/elastic/apm-server/blob/main/CONTRIBUTING.md for more tips on contributing.
-->

## Motivation/summary

Write to a byte buffer instead of writing to the gzip writer, then compress the payload once we are about to flush it to ES. Benchmarks shows a significant improvement in cpu time and switching to an optimized gzip library improves the situation even further.

NOTE: the NoCompression benchmark shows no improvements as this change only affect the bulk indexer with compression enabled.

Benchmarks:

```
name                                       old time/op    new time/op    delta
ModelIndexer/NoCompression-20                 685ns ± 2%     688ns ± 2%      ~     (p=0.279 n=8+8)
ModelIndexer/NoCompressionScaling-20          660ns ± 3%     645ns ± 5%      ~     (p=0.189 n=7+8)
ModelIndexer/BestSpeed-20                    1.40µs ± 2%    0.74µs ± 3%   -47.09%  (p=0.000 n=10+10)
ModelIndexer/BestSpeedScaling-20              832ns ± 7%     714ns ±12%   -14.15%  (p=0.000 n=10+10)
ModelIndexer/DefaultCompression-20           2.23µs ± 3%    0.78µs ± 3%   -65.23%  (p=0.000 n=10+10)
ModelIndexer/DefaultCompressionScaling-20    1.18µs ± 4%    0.66µs ± 3%   -44.10%  (p=0.000 n=10+9)
ModelIndexer/BestCompression-20              4.50µs ± 3%    0.94µs ± 1%   -79.07%  (p=0.000 n=10+9)
ModelIndexer/BestCompressionScaling-20       1.74µs ± 2%    0.96µs ± 3%   -44.70%  (p=0.000 n=10+10)

name                                       old speed      new speed      delta
ModelIndexer/NoCompression-20              3.62GB/s ± 2%  3.61GB/s ± 2%      ~     (p=0.279 n=8+8)
ModelIndexer/NoCompressionScaling-20       3.76GB/s ± 3%  3.85GB/s ± 5%      ~     (p=0.189 n=7+8)
ModelIndexer/BestSpeed-20                  1.78GB/s ± 2%  3.36GB/s ± 3%   +89.02%  (p=0.000 n=10+10)
ModelIndexer/BestSpeedScaling-20           2.99GB/s ± 7%  3.50GB/s ±13%   +16.92%  (p=0.000 n=10+10)
ModelIndexer/DefaultCompression-20         1.11GB/s ± 3%  3.20GB/s ± 3%  +187.64%  (p=0.000 n=10+10)
ModelIndexer/DefaultCompressionScaling-20  2.10GB/s ± 4%  3.76GB/s ± 3%   +78.81%  (p=0.000 n=10+9)
ModelIndexer/BestCompression-20             552MB/s ± 3%  2636MB/s ± 1%  +377.79%  (p=0.000 n=10+9)
ModelIndexer/BestCompressionScaling-20     1.42GB/s ± 2%  2.58GB/s ± 3%   +80.89%  (p=0.000 n=10+10)

name                                       old alloc/op   new alloc/op   delta
ModelIndexer/NoCompression-20                1.92kB ± 0%    1.92kB ± 0%      ~     (p=0.294 n=8+10)
ModelIndexer/NoCompressionScaling-20         1.92kB ± 0%    1.92kB ± 0%      ~     (p=0.106 n=9+10)
ModelIndexer/BestSpeed-20                    2.22kB ± 0%    1.98kB ± 0%   -10.49%  (p=0.000 n=9+10)
ModelIndexer/BestSpeedScaling-20             2.11kB ± 0%    2.02kB ± 0%    -4.33%  (p=0.000 n=10+9)
ModelIndexer/DefaultCompression-20           2.40kB ± 1%    1.99kB ± 0%   -17.26%  (p=0.000 n=10+10)
ModelIndexer/DefaultCompressionScaling-20    2.20kB ± 0%    2.02kB ± 0%    -8.09%  (p=0.000 n=8+10)
ModelIndexer/BestCompression-20              2.61kB ± 0%    2.00kB ± 0%   -23.43%  (p=0.000 n=10+10)
ModelIndexer/BestCompressionScaling-20       2.37kB ± 0%    1.99kB ± 0%   -15.87%  (p=0.000 n=9+9)

name                                       old allocs/op  new allocs/op  delta
ModelIndexer/NoCompression-20                  20.0 ± 0%      20.0 ± 0%      ~     (all equal)
ModelIndexer/NoCompressionScaling-20           20.0 ± 0%      20.0 ± 0%      ~     (all equal)
ModelIndexer/BestSpeed-20                      21.0 ± 0%      20.0 ± 0%    -4.76%  (p=0.000 n=10+10)
ModelIndexer/BestSpeedScaling-20               21.0 ± 0%      20.0 ± 0%    -4.76%  (p=0.000 n=10+10)
ModelIndexer/DefaultCompression-20             22.5 ± 2%      20.0 ± 0%   -11.11%  (p=0.000 n=10+10)
ModelIndexer/DefaultCompressionScaling-20      21.0 ± 0%      20.0 ± 0%    -4.76%  (p=0.000 n=10+10)
ModelIndexer/BestCompression-20                24.0 ± 0%      20.0 ± 0%   -16.67%  (p=0.000 n=10+10)
ModelIndexer/BestCompressionScaling-20         22.0 ± 0%      20.0 ± 0%    -9.09%  (p=0.000 n=10+10)
```

## Checklist

<!--
Delete irrelevant items. The changelog should only be updated for user-facing changes.
Once the PR is ready for review there should be no unticked boxes.
-->

- [ ] Update [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/main/CHANGELOG.asciidoc)
- [ ] Update [package changelog.yml](https://github.com/elastic/apm-server/blob/main/apmpackage/apm/changelog.yml) (only if changes to `apmpackage` have been made)
- [ ] Documentation has been updated

For functional changes, consider:
- Is it observable through the addition of either **logging** or **metrics**?
- Is its use being published in **telemetry** to enable product improvement?
- Have system tests been added to avoid regression?

## How to test these changes

<!--
Explain how this PR can be tested by the reviewer: commands, dependencies, steps, etc.
If it is self-explanatory, delete this section.
-->

## Related issues

<!--
Reference the related issue(s), and make use of magic keywords where it makes sense
https://help.github.com/articles/closing-issues-using-keywords/.
-->